AI healthcare Diagnosis final/AI healthcare Diagnosis (2)/AI healthcare Diagnosis/AI healthcare Diagnosis/requirements.txt
flask==2.0.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
tensorflow==2.6.0
Pillow==8.3.2
python-dotenv==0.19.0
opencv-python==4.5.3.56
pytesseract==0.3.8 